fix: move Windows-only wallpaper helpers to platform file, fix PS test
- Moved fetchWallpaperArchiveURLFromTap, downloadAndExtractWallpaperArchive,
  extractZip, extractZipEntry, extractTarGz to wallpapers_windows.go so they
  are only compiled on Windows (eliminating 'unused' lint errors on Linux)
- Fixed errcheck lint violations by wrapping Close() calls in func() { _ = ... }()
- Updated TestShellEnableForAllShells to check for '# bluefin-cli shell-config'
  in PowerShell profiles (the caching approach no longer writes 'bluefin-cli init'
  literally)
